+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Decrypted hashcat trophy files beautified
  - Added confirm questions before attacks if no target selected
  - WPA/WPA2 decrypted files using aircrack now have trophy file
+ - Some outputs colorized using ccze
 
 ### 7.1
  - Added default choice for some yes/no question. Regexp improved